client/requirements.txt
Flask==0.10.1

# Flask dependencies.
werkzeug==0.14.1
itsdangerous==1.1.0
markupsafe==1.1.1

# Jinja 2 is the HTML templating library that Flask uses.
jinja2==3.0.3

# Application server for both development and production.
gunicorn==19.4.5

# Testing and static analysis.
pytest==5.1.0
pytest-cov==2.7.1
flake8==3.7.8

# CLI.
Click==6.4

# gRPC and its tools
grpcio==1.46.3
grpcio-tools==1.46.3